Fighter Container is added
