+++ css/_weapon-pop-up.css
@@ -105,3 +105,17 @@
 		width: 4rem;
 	}
 }
+.fighter-pop-up {
+	position: absolute;
+	top: 50%;
+	left: 50%;
+	transform: translate(-50%, -50%);
+
+	z-index: 100;
+
+	/* display: flex;
+	flex-direction: row-reverse; */
+
+	height: 65vh;
+	/* align-items: flex-start; */
+}
